Database built from the ground-up - New entities and entity relations - Getting Draft for recyclerView and EditNoteFragment.kt works fine now
Known Issues
- DraftModelItems sometimes stored in incorrect order hence also retrieved in the same incorrect order

+++ app/src/main/java/com/example/notesapp/database/entities/Note.kt
@@ -10,9 +10,8 @@ import kotlinx.parcelize.Parcelize
 @Entity(tableName = "notes_table")
 data class Note(
     @PrimaryKey(autoGenerate = true)
-    var ID: Int,
+    var ID: Long,
     var noteText: String,
     @Nullable
     var noteImageBannerURL: String?,
-//    var draftID: Long
 ) : Parcelable
